Struct cargo_arch::config::arch::ArchConfig

pub struct ArchConfig {
    pub maintainers: Vec<String>,
    pub pkgname: String,
    pub pkgver: String,
    pub pkgrel: String,
    pub epoch: String,
    pub pkgdesc: String,
    pub url: String,
    pub license: Vec<String>,
    pub install: String,
    pub changelog: String,
    pub source: Vec<String>,
    pub validpgpkeys: Vec<String>,
    pub noextract: Vec<String>,
    pub md5sums: Vec<String>,
    pub sha1sums: Vec<String>,
    pub sha256sums: Vec<String>,
    pub sha384sums: Vec<String>,
    pub sha512sums: Vec<String>,
    pub groups: Vec<String>,
    pub arch: Vec<String>,
    pub backup: Vec<String>,
    pub depends: Vec<String>,
    pub makedepends: Vec<String>,
    pub checkdepends: Vec<String>,
    pub optdepends: Vec<String>,
    pub conflicts: Vec<String>,
    pub provides: Vec<String>,
    pub replaces: Vec<String>,
    pub options: Vec<String>,
}

see man PKGBUILD and https://wiki.archlinux.org/index.php/PKGBUILD

Fields
maintainers: Vec<String>

The maintainers of the package

pkgname: String

The name of the package.

pkgver: String

The version of the software as released from the author.

pkgrel: String

This is the release number specific to the Arch Linux release.

epoch: String

Used to force the package to be seen as newer than any previous versions with a lower epoch, even if the version number would normally not trigger such an upgrade.

pkgdesc: String

This should be a brief description of the package and its functionality.

url: String

This field contains a URL that is associated with the software being packaged. This is typically the project’s web site.

license: Vec<String>

This field specifies the license(s) that apply to the package.

install: String

Specifies a special install script that is to be included in the package.

changelog: String

Specifies a changelog file that is to be included in the package.

source: Vec<String>

An array of source files required to build the package.

validpgpkeys: Vec<String>

An array of PGP fingerprints.

noextract: Vec<String>

An array of file names corresponding to those from the source array.

md5sums: Vec<String>

This array contains an MD5 hash for every source file specified in the source array (in the same order).

sha1sums: Vec<String>

Alternative integrity checks that makepkg supports; these all behave similar to the md5sums option described above.

sha256sums: Vec<String>

Alternative integrity checks that makepkg supports; these all behave similar to the md5sums option described above.

sha384sums: Vec<String>

Alternative integrity checks that makepkg supports; these all behave similar to the md5sums option described above.

sha512sums: Vec<String>

Alternative integrity checks that makepkg supports; these all behave similar to the md5sums option described above.

groups: Vec<String>

An array of symbolic names that represent groups of packages, allowing you to install multiple packages by requesting a single target.

arch: Vec<String>

Defines on which architectures the given package is available.

backup: Vec<String>

An array of file names, without preceding slashes, that should be backed up if the package is removed or upgraded.

depends: Vec<String>

An array of packages this package depends on to run.

makedepends: Vec<String>

An array of packages this package depends on to build but are not needed at runtime.

checkdepends: Vec<String>

An array of packages this package depends on to run its test suite but are not needed at runtime.

optdepends: Vec<String>

An array of packages (and accompanying reasons) that are not essential for base functionality, but may be necessary to make full use of the contents of this package.

conflicts: Vec<String>

An array of packages that will conflict with this package.

provides: Vec<String>

An array of "virtual provisions" this package provides.

replaces: Vec<String>

An array of packages this package should replace.

options: Vec<String>

This array allows you to override some of makepkg’s default behavior when building packages.

Methods

impl ArchConfig

fn new() -> ArchConfig

fn generate_pkgbuild(&self)