Struct cargo_arch::config::arch::CargoArch

pub struct CargoArch {
    pub maintainers: Option<Vec<String>>,
    pub pkgname: Option<String>,
    pub pkgver: Option<String>,
    pub pkgrel: Option<String>,
    pub epoch: Option<String>,
    pub pkgdesc: Option<String>,
    pub url: Option<String>,
    pub license: Option<Vec<String>>,
    pub install: Option<String>,
    pub changelog: Option<String>,
    pub source: Option<Vec<String>>,
    pub validpgpkeys: Option<Vec<String>>,
    pub noextract: Option<Vec<String>>,
    pub md5sums: Option<Vec<String>>,
    pub sha1sums: Option<Vec<String>>,
    pub sha256sums: Option<Vec<String>>,
    pub sha384sums: Option<Vec<String>>,
    pub sha512sums: Option<Vec<String>>,
    pub groups: Option<Vec<String>>,
    pub arch: Option<Vec<String>>,
    pub backup: Option<Vec<String>>,
    pub depends: Option<Vec<String>>,
    pub makedepends: Option<Vec<String>>,
    pub checkdepends: Option<Vec<String>>,
    pub optdepends: Option<Vec<String>>,
    pub conflicts: Option<Vec<String>>,
    pub provides: Option<Vec<String>>,
    pub replaces: Option<Vec<String>>,
    pub options: Option<Vec<String>>,
}

data in [package.metadata.arch]

Fields
maintainers: Option<Vec<String>>

The maintainers of the package

pkgname: Option<String>

The name of the package.

pkgver: Option<String>

The version of the software as released from the author.

pkgrel: Option<String>

This is the release number specific to the Arch Linux release.

epoch: Option<String>

Used to force the package to be seen as newer than any previous versions with a lower epoch, even if the version number would normally not trigger such an upgrade.

pkgdesc: Option<String>

This should be a brief description of the package and its functionality.

url: Option<String>

This field contains a URL that is associated with the software being packaged. This is typically the project’s web site.

license: Option<Vec<String>>

This field specifies the license(s) that apply to the package.

install: Option<String>

Specifies a special install script that is to be included in the package.

changelog: Option<String>

Specifies a changelog file that is to be included in the package.

source: Option<Vec<String>>

An array of source files required to build the package.

validpgpkeys: Option<Vec<String>>

An array of PGP fingerprints.

noextract: Option<Vec<String>>

An array of file names corresponding to those from the source array.

md5sums: Option<Vec<String>>

This array contains an MD5 hash for every source file specified in the source array (in the same order).

sha1sums: Option<Vec<String>>

Alternative integrity checks that makepkg supports; these all behave similar to the md5sums option described above.

sha256sums: Option<Vec<String>>

Alternative integrity checks that makepkg supports; these all behave similar to the md5sums option described above.

sha384sums: Option<Vec<String>>

Alternative integrity checks that makepkg supports; these all behave similar to the md5sums option described above.

sha512sums: Option<Vec<String>>

Alternative integrity checks that makepkg supports; these all behave similar to the md5sums option described above.

groups: Option<Vec<String>>

An array of symbolic names that represent groups of packages, allowing you to install multiple packages by requesting a single target.

arch: Option<Vec<String>>

Defines on which architectures the given package is available.

backup: Option<Vec<String>>

An array of file names, without preceding slashes, that should be backed up if the package is removed or upgraded.

depends: Option<Vec<String>>

An array of packages this package depends on to run.

makedepends: Option<Vec<String>>

An array of packages this package depends on to build but are not needed at runtime.

checkdepends: Option<Vec<String>>

An array of packages this package depends on to run its test suite but are not needed at runtime.

optdepends: Option<Vec<String>>

An array of packages (and accompanying reasons) that are not essential for base functionality, but may be necessary to make full use of the contents of this package.

conflicts: Option<Vec<String>>

An array of packages that will conflict with this package.

provides: Option<Vec<String>>

An array of "virtual provisions" this package provides.

replaces: Option<Vec<String>>

An array of packages this package should replace.

options: Option<Vec<String>>

This array allows you to override some of makepkg’s default behavior when building packages.